{{cookiecutter.github_repository}}/requirements/development.txt
-r common.txt

# Documentation
# -------------------------------------
https://github.com/r0wb0t/markdown-urlize/archive/8bf9bd75cf61ba255bd91a9e808add925d95412e.zip
https://github.com/tomchristie/mkdocs/archive/master.zip
isort==4.2.1

# Debugging
# -------------------------------------
django-debug-toolbar==1.3.2
ipdb==0.8.1
ipython==4.0.0

# Testing and coverage
# -------------------------------------
coverage==4.0
factory_boy==2.5.2
flake8==2.4.1
git+git://github.com/mverteuil/pytest-ipdb.git
mock==1.3.0
pytest-django==2.8.0
pytest-pythonpath==0.7

# Versioning
# -------------------------------------
bumpversion==0.5.3
